moved UserSchema to user model, created PostSchema
